requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-chroma
langchainhub
langchain-experimental
sentence-transformers
transformers
torch
chromadb
pypdf
rapidocr-onnxruntime
